refactor (LeagueUtil) split off ChampionUtil
